Sketch edit of hohmann_moon.svg
hohmann_moon.svg was originally created in Inkscape back in 2015.
However now I couldn't get Inkscape to run at all. It asked for a Quartz path and then simply didn't work. I also observed in the Finder prompt that asked for Quartz.app that it was very ugly on retina display.

So I decided to try Sketch instead. I needed to make a small change in hohmann_moon.svg, and imported it into Sketch. The following changes were made:
1. Change the direction of the orbit arond the moon in the zoombox.
2. Enlarged the whole figure to be able to position stuff more easily on a 0.5 px level.
3. Clean up the figure a bit.